requirements.txt
flask==1.1.1
flask_cors==3.0.8
flask_limiter==1.0.1
flask_login==0.4.1
flask_restful==0.3.7
flask_sqlalchemy==2.4.1
kazoo==2.6.1
pymaid==0.4.0b5
redis==3.3.8
requests==2.22.0
sqlalchemy==1.3.8
pymysql==0.9.3
celery==4.3.0
pyopenssl==19.0.0
oss2==2.8.0
uWSGI==2.0.18

